+++ src/components/DailyWorkers/DailyWorkersStyle.tsx
