+++ src/pages/admin/MainPage/ScheduleSection/components/DailyWorkers.tsx
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { getDailyWorkers } from 'apis/getSchedule';
-import DailyWorkersTemplate from 'components/@commons-feature/calendar/DailyWorkersTemplate';
+import DailyWorkersTemplate from 'components/DailyWorkers/DailyWorkersTemplate';
 import FlexContainer from 'components/@commons/FlexContainer';
 import React from 'react';
